requirements.txt
requests
PyQt6
Flask==3.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
pytz==2024.1
python-dotenv==1.0.1
Flask-CORS==4.0.0